pyproject.toml
[project]
name = "web-automation-framework"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pytest>=8.0.0",
  "pytest-xdist>=3.5.0",
  "python-dotenv>=1.0.0",
  "playwright>=1.46.0",
]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
